third_party/WebKit/LayoutTests/fast/forms/date/date-setrangetext.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../js/resources/js-test-pre.js"></script>
<script src="../resources/common-setrangetext.js"></script>
</head>
<body>
<script>
description("Test setRangeText() method is not available in date inputs.");
runTestsShouldFail("input", { type: "date" });
</script>
</body>
</html>
