third_party/WebKit/LayoutTests/fast/forms/date/date-setrangetext-expected.txt

Test setRangeText() method is not available in date inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Running tests on input with attributes: {"type":"date"}

element.value = '0123456789XYZ'
PASS element.setRangeText('ABC', 0, 0) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS element.value is ""
PASS successfullyParsed is true

TEST COMPLETE
